Update pnSelf.exe
Now with reply function, disappearing message function, improved spam function, improved dmsg function, fancier help function, and wikipedia search function. Fixed issue with old ass discord.js library which is beyond deprecated but is required for the selfbot to function. Ill have to look into a better solution for this in the future.